requirements.txt
annotated-types
anyio
certifi
click
colorama
dnspython
email_validator
fastapi
fastapi-cli
h11
httpcore
httptools
httpx
idna
Jinja2
markdown-it-py
MarkupSafe
mdurl
pydantic
pydantic_core
Pygments
pypdf
python-dotenv
python-multipart
PyYAML
rich
rich-toolkit
shellingham
sniffio
starlette
typer
typing-inspection
typing_extensions
uvicorn
watchfiles
websockets
langchain-openai
pinecone
pinecone-plugin-interface
langchain_core